AI now moves towards an item automatically

+++ ZombieGame/project/Graph/Dijkstra.cpp
@@ -20,7 +20,7 @@ std::vector<GraphNode*> Dijkstra::FindPath(Graph* pGraph, GraphNode* pStartNode,
 	{
 		
 		bool endNodeFound = false;
-		for (int i = 0; i < openList.size(); ++i)
+		for (size_t i = 0; i < openList.size(); ++i)
 		{
 			auto currentRecord = lastRecord = openList[i];
